osf-ml-predictor/requirements.txt
fastapi>=0.109.0
uvicorn>=0.25.0
neo4j>=5.15.0
autogluon.timeseries>=1.0.0
autogluon.tabular>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
httpx>=0.26.0
pydantic>=2.0.0
